Add the utililty custom hook to get the page badge based on current view
